+++ bindings/java/src/test/com/apple/foundationdb/test/StackTester.java
@@ -434,13 +434,14 @@ public class StackTester {
 						db.options().setMaxWatches(10001);
 						db.options().setDatacenterId("dc_id");
 						db.options().setMachineId("machine_id");
+						db.options().setSnapshotRywEnable();
+						db.options().setSnapshotRywDisable();
+						db.options().setTransactionLoggingMaxFieldLength(1000);
 						db.options().setTransactionTimeout(100000);
 						db.options().setTransactionTimeout(0);
 						db.options().setTransactionMaxRetryDelay(100);
 						db.options().setTransactionRetryLimit(10);
 						db.options().setTransactionRetryLimit(-1);
-						db.options().setSnapshotRywEnable();
-						db.options().setSnapshotRywDisable();
 
 						tr.options().setPrioritySystemImmediate();
 						tr.options().setPriorityBatch();
@@ -449,6 +450,7 @@ public class StackTester {
 						tr.options().setReadYourWritesDisable();
 						tr.options().setReadSystemKeys();
 						tr.options().setAccessSystemKeys();
+						tr.options().setTransactionLoggingMaxFieldLength(1000);
 						tr.options().setTimeout(60*1000);
 						tr.options().setRetryLimit(50);
 						tr.options().setMaxRetryDelay(100);